줄바꿈 처리

+++ hallabong(team1)/src/main/java/com/hallabong/qna/vo/QnaVO.java
@@ -7,15 +7,15 @@ import lombok.Data;
 @Data
 public class QnaVO {
 
-	private long no;
+	private Integer no;
 	private String title;
 	private String content; 
 	private String id;
 	private String name;
 	private Date writeDate;
 	private long cnt;
-	private long refNo;
+	private Integer refNo;
 	private long ordNo;
 	private long levNo;
-	private long parentNo;
+	private Integer parentNo;
 }
